CHanged getAppId from returning long to int

+++ src/main/java/com/steamrankings/service/api/games/SteamGame.java
@@ -31,7 +31,7 @@ public class SteamGame {
     }
 
     @JsonIgnore
-    public long getAppId() {
+    public int getAppId() {
         return appid;
     }
